+++ src/containers/modal/feature/feature.jsx
@@ -15,7 +15,7 @@ import {
 } from '@chakra-ui/react'
 import { ExternalLinkIcon } from '@chakra-ui/icons'
 
-const version = "v0.8"
+const version = "v0.9"
 
 export default function Feature(props) {
 
@@ -24,6 +24,7 @@ export default function Feature(props) {
     return (
         <>
             <Button
+                className="version-btn"
                 color={props.colorMode === "light" ? "#1941b3" : "white"} onClick={onOpen}>
                 {version}
             </Button>
@@ -67,6 +68,18 @@ export default function Feature(props) {
                                 <UnorderedList>
                                     <ListItem>
                                         Invoice Printer is serverless now ! 🎉🎉
+                                    </ListItem>
+                                </UnorderedList>
+                                <ListItem>v0.9</ListItem>
+                                <UnorderedList>
+                                    <ListItem>
+                                        Added Logo Input
+                                    </ListItem>
+                                    <ListItem>
+                                        Edited mobile designed
+                                    </ListItem>
+                                    <ListItem>
+                                        Removed minimize button
                                     </ListItem>
                                 </UnorderedList>
                             </UnorderedList>